+++ programs/drift/src/instructions/admin.rs
@@ -11,6 +11,7 @@ use solana_program::msg;
 
 use crate::controller::token::close_vault;
 use crate::error::ErrorCode;
+use crate::ids::fuel_airdrop_wallet;
 use crate::instructions::constraints::*;
 use crate::math::casting::Cast;
 use crate::math::constants::{
@@ -39,6 +40,7 @@ use crate::state::oracle::{
     HistoricalIndexData, HistoricalOracleData, OraclePriceData, OracleSource, PrelaunchOracle,
     PrelaunchOracleParams,
 };
+use crate::state::oracle_map::OracleMap;
 use crate::state::paused_operations::{InsuranceFundOperation, PerpOperation, SpotOperation};
 use crate::state::perp_market::{
     ContractTier, ContractType, InsuranceClaim, MarketStatus, PerpMarket, PoolBalance, AMM,
@@ -48,7 +50,7 @@ use crate::state::spot_market::{
 };
 use crate::state::state::{ExchangeStatus, FeeStructure, OracleGuardRails, State};
 use crate::state::traits::Size;
-use crate::state::user::UserStats;
+use crate::state::user::{User, UserStats};
 use crate::validate;
 use crate::validation::fee_structure::validate_fee_structure;
 use crate::validation::margin::{validate_margin, validate_margin_weights};
@@ -134,6 +136,8 @@ pub fn handle_initialize_spot_market(
 
     let spot_market_index = get_then_update_id!(state, number_of_spot_markets);
 
+    msg!("initializing spot market {}", spot_market_index);
+
     if oracle_source == OracleSource::QuoteAsset {
         // catches inconsistent parameters
         validate!(
@@ -147,6 +151,8 @@ pub fn handle_initialize_spot_market(
             ErrorCode::InvalidSpotMarketInitialization,
             "For OracleSource::QuoteAsset, spot_market_index must be QUOTE_SPOT_MARKET_INDEX"
         )?;
+    } else {
+        OracleMap::validate_oracle_account_info(&ctx.accounts.oracle)?;
     }
 
     let oracle_price_data = get_oracle_price(
@@ -285,7 +291,8 @@ pub fn handle_initialize_spot_market(
         fuel_boost_borrows: 0,
         fuel_boost_taker: 0,
         fuel_boost_maker: 0,
-        padding: [0; 43],
+        fuel_boost_insurance: 0,
+        padding: [0; 42],
         insurance_fund: InsuranceFund {
             vault: *ctx.accounts.insurance_fund_vault.to_account_info().key,
             unstaking_period: THIRTEEN_DAY,
@@ -532,6 +539,7 @@ pub fn handle_initialize_perp_market(
     amm_jit_intensity: u8,
     name: [u8; 32],
 ) -> Result<()> {
+    msg!("perp market {}", market_index);
     let perp_market_pubkey = ctx.accounts.perp_market.to_account_info().key;
     let perp_market = &mut ctx.accounts.perp_market.load_init()?;
     let clock = Clock::get()?;
@@ -570,6 +578,8 @@ pub fn handle_initialize_perp_market(
 
     let (min_base_asset_reserve, max_base_asset_reserve) =
         amm::calculate_bid_ask_bounds(concentration_coef, amm_base_asset_reserve)?;
+
+    OracleMap::validate_oracle_account_info(&ctx.accounts.oracle)?;
 
     // Verify oracle is readable
     let (oracle_price, oracle_delay, last_oracle_price_twap) = match oracle_source {
@@ -850,6 +860,7 @@ pub fn handle_delete_initialized_perp_market(
     market_index: u16,
 ) -> Result<()> {
     let perp_market = &mut ctx.accounts.perp_market.load()?;
+    msg!("perp market {}", perp_market.market_index);
     let state = &mut ctx.accounts.state;
 
     // to preserve all protocol invariants, can only remove the last market if it hasn't been "activated"
@@ -890,6 +901,7 @@ pub fn handle_delete_initialized_spot_market(
     market_index: u16,
 ) -> Result<()> {
     let spot_market = ctx.accounts.spot_market.load()?;
+    msg!("spot market {}", spot_market.market_index);
     let state = &mut ctx.accounts.state;
 
     // to preserve all protocol invariants, can only remove the last market if it hasn't been "activated"
@@ -972,7 +984,18 @@ pub fn handle_update_spot_market_oracle(
     oracle_source: OracleSource,
 ) -> Result<()> {
     let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
+    msg!("updating spot market {} oracle", spot_market.market_index);
     let clock = Clock::get()?;
+
+    OracleMap::validate_oracle_account_info(&ctx.accounts.oracle)?;
+
+    validate!(
+        ctx.accounts.oracle.key == &oracle,
+        ErrorCode::DefaultError,
+        "oracle account info ({:?}) and ix data ({:?}) must match",
+        ctx.accounts.oracle.key,
+        oracle
+    )?;
 
     // Verify oracle is readable
     let OraclePriceData {
@@ -1006,6 +1029,7 @@ pub fn handle_update_spot_market_expiry(
     expiry_ts: i64,
 ) -> Result<()> {
     let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
+    msg!("updating spot market {} expiry", spot_market.market_index);
     let now = Clock::get()?.unix_timestamp;
 
     validate!(
@@ -1032,6 +1056,79 @@ pub fn handle_update_spot_market_expiry(
     Ok(())
 }
 
+pub fn handle_init_user_fuel(
+    ctx: Context<InitUserFuel>,
+    fuel_bonus_deposits: Option<u32>,
+    fuel_bonus_borrows: Option<u32>,
+    fuel_bonus_taker: Option<u32>,
+    fuel_bonus_maker: Option<u32>,
+    fuel_bonus_insurance: Option<u32>,
+) -> Result<()> {
+    let clock: Clock = Clock::get()?;
+    let now_u32 = clock.unix_timestamp as u32;
+
+    let user = &mut load_mut!(ctx.accounts.user)?;
+    let user_stats = &mut load_mut!(ctx.accounts.user_stats)?;
+
+    validate!(
+        user.last_fuel_bonus_update_ts == 0,
+        ErrorCode::DefaultError,
+        "User must not have begun earning fuel"
+    )?;
+
+    if let Some(fuel_bonus_deposits) = fuel_bonus_deposits {
+        msg!(
+            "user_stats.fuel_deposits {:?} -> {:?}",
+            user_stats.fuel_deposits,
+            user_stats.fuel_deposits.saturating_add(fuel_bonus_deposits)
+        );
+        user_stats.fuel_deposits = user_stats.fuel_deposits.saturating_add(fuel_bonus_deposits);
+    }
+    if let Some(fuel_bonus_borrows) = fuel_bonus_borrows {
+        msg!(
+            "user_stats.fuel_borrows {:?} -> {:?}",
+            user_stats.fuel_borrows,
+            user_stats.fuel_borrows.saturating_add(fuel_bonus_borrows)
+        );
+        user_stats.fuel_borrows = user_stats.fuel_borrows.saturating_add(fuel_bonus_borrows);
+    }
+
+    if let Some(fuel_bonus_taker) = fuel_bonus_taker {
+        msg!(
+            "user_stats.fuel_taker {:?} -> {:?}",
+            user_stats.fuel_taker,
+            user_stats.fuel_taker.saturating_add(fuel_bonus_taker)
+        );
+        user_stats.fuel_taker = user_stats.fuel_taker.saturating_add(fuel_bonus_taker);
+    }
+    if let Some(fuel_bonus_maker) = fuel_bonus_maker {
+        msg!(
+            "user_stats.fuel_maker {:?} -> {:?}",
+            user_stats.fuel_maker,
+            user_stats.fuel_maker.saturating_add(fuel_bonus_maker)
+        );
+        user_stats.fuel_maker = user_stats.fuel_maker.saturating_add(fuel_bonus_maker);
+    }
+
+    if let Some(fuel_bonus_insurance) = fuel_bonus_insurance {
+        msg!(
+            "user_stats.fuel_insurance {:?} -> {:?}",
+            user_stats.fuel_insurance,
+            user_stats
+                .fuel_insurance
+                .saturating_add(fuel_bonus_insurance)
+        );
+        user_stats.fuel_insurance = user_stats
+            .fuel_insurance
+            .saturating_add(fuel_bonus_insurance);
+    }
+
+    user.last_fuel_bonus_update_ts = now_u32;
+    user_stats.last_fuel_if_bonus_update_ts = now_u32;
+
+    Ok(())
+}
+
 #[access_control(
     perp_market_valid(&ctx.accounts.perp_market)
 )]
@@ -1039,8 +1136,10 @@ pub fn handle_update_perp_market_expiry(
     ctx: Context<AdminUpdatePerpMarket>,
     expiry_ts: i64,
 ) -> Result<()> {
-    let clock = Clock::get()?;
+    let clock: Clock = Clock::get()?;
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("updating perp market {} expiry", perp_market.market_index);
+
     validate!(
         clock.unix_timestamp < expiry_ts,
         ErrorCode::DefaultError,
@@ -1075,6 +1174,11 @@ pub fn handle_move_amm_price(
     sqrt_k: u128,
 ) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+
+    msg!(
+        "moving amm price for perp market {}",
+        perp_market.market_index
+    );
 
     let base_asset_reserve_before = perp_market.amm.base_asset_reserve;
     let quote_asset_reserve_before = perp_market.amm.quote_asset_reserve;
@@ -1134,6 +1238,11 @@ pub fn handle_recenter_perp_market_amm(
     sqrt_k: u128,
 ) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+
+    msg!(
+        "recentering amm for perp market {}",
+        perp_market.market_index
+    );
 
     let base_asset_reserve_before = perp_market.amm.base_asset_reserve;
     let quote_asset_reserve_before = perp_market.amm.quote_asset_reserve;
@@ -1205,6 +1314,16 @@ pub fn handle_update_perp_market_amm_summary_stats(
 ) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
     let spot_market = &mut load!(ctx.accounts.spot_market)?;
+
+    msg!(
+        "updating amm summary stats for perp market {}",
+        perp_market.market_index
+    );
+
+    msg!(
+        "updating amm summary stats for spot market {}",
+        spot_market.market_index
+    );
 
     let clock = Clock::get()?;
     let price_oracle = &ctx.accounts.oracle;
@@ -1286,6 +1405,16 @@ pub fn handle_settle_expired_market_pools_to_revenue_pool(
     let spot_market: &mut std::cell::RefMut<'_, SpotMarket> =
         &mut load_mut!(ctx.accounts.spot_market)?;
     let state = &ctx.accounts.state;
+
+    msg!(
+        "settling expired market pools to revenue pool for perp market {}",
+        perp_market.market_index
+    );
+
+    msg!(
+        "settling expired market pools to revenue pool for spot market {}",
+        spot_market.market_index
+    );
 
     let clock = Clock::get()?;
     let now = clock.unix_timestamp;
@@ -1397,6 +1526,12 @@ pub fn handle_deposit_into_perp_market_fee_pool(
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
 
     msg!(
+        "depositing {} into perp market {} fee pool",
+        amount,
+        perp_market.market_index
+    );
+
+    msg!(
         "perp_market.amm.total_fee_minus_distributions: {:?} -> {:?}",
         perp_market.amm.total_fee_minus_distributions,
         perp_market
@@ -1441,6 +1576,11 @@ pub fn handle_repeg_amm_curve(ctx: Context<RepegCurve>, new_peg_candidate: u128)
     let clock_slot = clock.slot;
 
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!(
+        "repegging amm curve for perp market {}",
+        perp_market.market_index
+    );
+
     let price_oracle = &ctx.accounts.oracle;
     let OraclePriceData {
         price: oracle_price,
@@ -1529,6 +1669,10 @@ pub fn handle_update_amm_oracle_twap(ctx: Context<RepegCurve>) -> Result<()> {
     let now = clock.unix_timestamp;
 
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!(
+        "updating amm oracle twap for perp market {}",
+        perp_market.market_index
+    );
     let price_oracle = &ctx.accounts.oracle;
     let oracle_twap = perp_market.amm.get_oracle_twap(price_oracle, clock.slot)?;
 
@@ -1622,6 +1766,7 @@ pub fn handle_update_k(ctx: Context<AdminUpdateK>, sqrt_k: u128) -> Result<()> {
 
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
 
+    msg!("updating k for perp market {}", perp_market.market_index);
     let base_asset_amount_long = perp_market.amm.base_asset_amount_long.unsigned_abs();
     let base_asset_amount_short = perp_market.amm.base_asset_amount_short.unsigned_abs();
     let base_asset_amount_with_amm = perp_market.amm.base_asset_amount_with_amm;
@@ -1818,6 +1963,10 @@ pub fn handle_reset_amm_oracle_twap(ctx: Context<RepegCurve>) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
 
     msg!(
+        "resetting amm oracle twap for perp market {}",
+        perp_market.market_index
+    );
+    msg!(
         "perp_market.amm.historical_oracle_data.last_oracle_price_twap: {:?} -> {:?}",
         perp_market
             .amm
@@ -1856,6 +2005,12 @@ pub fn handle_update_perp_market_margin_ratio(
     margin_ratio_maintenance: u32,
 ) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+
+    msg!(
+        "updating perp market {} margin ratio",
+        perp_market.market_index
+    );
+
     validate_margin(
         margin_ratio_initial,
         margin_ratio_maintenance,
@@ -1889,6 +2044,11 @@ pub fn handle_update_perp_market_funding_period(
 ) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
 
+    msg!(
+        "updating funding period for perp market {}",
+        perp_market.market_index
+    );
+
     validate!(funding_period >= 0, ErrorCode::DefaultError)?;
 
     msg!(
@@ -1911,6 +2071,11 @@ pub fn handle_update_perp_market_max_imbalances(
     quote_max_insurance: u64,
 ) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+
+    msg!(
+        "updating perp market {} max imbalances",
+        perp_market.market_index
+    );
 
     let max_insurance_for_tier = match perp_market.contract_tier {
         ContractTier::A => INSURANCE_A_MAX,
@@ -2001,6 +2166,12 @@ pub fn handle_update_perp_liquidation_fee(
     if_liquidation_fee: u32,
 ) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+
+    msg!(
+        "updating perp market {} liquidation fee",
+        perp_market.market_index
+    );
+
     validate!(
         liquidator_fee.safe_add(if_liquidation_fee)? < LIQUIDATION_FEE_PRECISION,
         ErrorCode::DefaultError,
@@ -2046,6 +2217,7 @@ pub fn handle_update_insurance_fund_unstaking_period(
 ) -> Result<()> {
     let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
 
+    msg!("updating spot market {} IF unstaking period");
     msg!(
         "spot_market.insurance_fund.unstaking_period: {:?} -> {:?}",
         spot_market.insurance_fund.unstaking_period,
@@ -2065,6 +2237,11 @@ pub fn handle_update_spot_market_liquidation_fee(
     if_liquidation_fee: u32,
 ) -> Result<()> {
     let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
+    msg!(
+        "updating spot market {} liquidation fee",
+        spot_market.market_index
+    );
+
     validate!(
         liquidator_fee.safe_add(if_liquidation_fee)? < LIQUIDATION_FEE_PRECISION,
         ErrorCode::DefaultError,
@@ -2103,6 +2280,11 @@ pub fn handle_update_withdraw_guard_threshold(
 ) -> Result<()> {
     let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
     msg!(
+        "updating spot market withdraw guard threshold {}",
+        spot_market.market_index
+    );
+
+    msg!(
         "spot_market.withdraw_guard_threshold: {:?} -> {:?}",
         spot_market.withdraw_guard_threshold,
         withdraw_guard_threshold
@@ -2121,6 +2303,8 @@ pub fn handle_update_spot_market_if_factor(
     total_if_factor: u32,
 ) -> Result<()> {
     let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
+
+    msg!("spot market {}", spot_market.market_index);
 
     validate!(
         spot_market.market_index == spot_market_index,
@@ -2165,6 +2349,8 @@ pub fn handle_update_spot_market_revenue_settle_period(
     revenue_settle_period: i64,
 ) -> Result<()> {
     let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
+    msg!("spot market {}", spot_market.market_index);
+
     validate!(revenue_settle_period > 0, ErrorCode::DefaultError)?;
     msg!(
         "spot_market.revenue_settle_period: {:?} -> {:?}",
@@ -2184,6 +2370,7 @@ pub fn handle_update_spot_market_status(
 ) -> Result<()> {
     status.validate_not_deprecated()?;
     let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
+    msg!("spot market {}", spot_market.market_index);
 
     msg!(
         "spot_market.status: {:?} -> {:?}",
@@ -2203,6 +2390,8 @@ pub fn handle_update_spot_market_paused_operations(
     paused_operations: u8,
 ) -> Result<()> {
     let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
+    msg!("spot market {}", spot_market.market_index);
+
     spot_market.paused_operations = paused_operations;
 
     SpotOperation::log_all_operations_paused(spot_market.paused_operations);
@@ -2218,6 +2407,7 @@ pub fn handle_update_spot_market_asset_tier(
     asset_tier: AssetTier,
 ) -> Result<()> {
     let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
+    msg!("spot market {}", spot_market.market_index);
 
     if spot_market.initial_asset_weight > 0 {
         validate!(
@@ -2249,6 +2439,7 @@ pub fn handle_update_spot_market_margin_weights(
     imf_factor: u32,
 ) -> Result<()> {
     let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
+    msg!("spot market {}", spot_market.market_index);
 
     validate_margin_weights(
         spot_market.market_index,
@@ -2309,6 +2500,8 @@ pub fn handle_update_spot_market_borrow_rate(
     min_borrow_rate: Option<u8>,
 ) -> Result<()> {
     let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
+    msg!("spot market {}", spot_market.market_index);
+
     validate_borrow_rate(
         optimal_utilization,
         optimal_borrow_rate,
@@ -2361,6 +2554,7 @@ pub fn handle_update_spot_market_max_token_deposits(
     max_token_deposits: u64,
 ) -> Result<()> {
     let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
+    msg!("spot market {}", spot_market.market_index);
 
     msg!(
         "spot_market.max_token_deposits: {:?} -> {:?}",
@@ -2380,6 +2574,7 @@ pub fn handle_update_spot_market_max_token_borrows(
     max_token_borrows_fraction: u16,
 ) -> Result<()> {
     let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
+    msg!("spot market {}", spot_market.market_index);
 
     msg!(
         "spot_market.max_token_borrows_fraction: {:?} -> {:?}",
@@ -2413,6 +2608,7 @@ pub fn handle_update_spot_market_scale_initial_asset_weight_start(
     scale_initial_asset_weight_start: u64,
 ) -> Result<()> {
     let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
+    msg!("spot market {}", spot_market.market_index);
 
     msg!(
         "spot_market.scale_initial_asset_weight_start: {:?} -> {:?}",
@@ -2432,6 +2628,7 @@ pub fn handle_update_spot_market_orders_enabled(
     orders_enabled: bool,
 ) -> Result<()> {
     let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
+    msg!("spot market {}", spot_market.market_index);
 
     msg!(
         "spot_market.orders_enabled: {:?} -> {:?}",
@@ -2474,6 +2671,8 @@ pub fn handle_update_perp_market_status(
 
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
 
+    msg!("perp market {}", perp_market.market_index);
+
     msg!(
         "perp_market.status: {:?} -> {:?}",
         perp_market.status,
@@ -2492,6 +2691,8 @@ pub fn handle_update_perp_market_paused_operations(
     paused_operations: u8,
 ) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
+
     perp_market.paused_operations = paused_operations;
 
     PerpOperation::log_all_operations_paused(perp_market.paused_operations);
@@ -2507,6 +2708,7 @@ pub fn handle_update_perp_market_contract_tier(
     contract_tier: ContractTier,
 ) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
 
     msg!(
         "perp_market.contract_tier: {:?} -> {:?}",
@@ -2537,6 +2739,7 @@ pub fn handle_update_perp_market_imf_factor(
         "invalid unrealized pnl imf factor",
     )?;
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
 
     msg!(
         "perp_market.imf_factor: {:?} -> {:?}",
@@ -2579,6 +2782,7 @@ pub fn handle_update_perp_market_unrealized_asset_weight(
         "must enforce unrealized_initial_asset_weight <= unrealized_maintenance_asset_weight",
     )?;
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
 
     msg!(
         "perp_market.unrealized_initial_asset_weight: {:?} -> {:?}",
@@ -2611,6 +2815,8 @@ pub fn handle_update_perp_market_concentration_coef(
     )?;
 
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
+
     let prev_concentration_coef = perp_market.amm.concentration_coef;
     controller::amm::update_concentration_coef(&mut perp_market.amm, concentration_scale)?;
     let new_concentration_coef = perp_market.amm.concentration_coef;
@@ -2639,6 +2845,7 @@ pub fn handle_update_perp_market_curve_update_intensity(
         "invalid curve_update_intensity",
     )?;
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
 
     msg!(
         "perp_market.amm.curve_update_intensity: {} -> {}",
@@ -2658,6 +2865,7 @@ pub fn handle_update_perp_market_target_base_asset_amount_per_lp(
     target_base_asset_amount_per_lp: i32,
 ) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
 
     msg!(
         "perp_market.amm.target_base_asset_amount_per_lp: {} -> {}",
@@ -2674,6 +2882,7 @@ pub fn handle_update_perp_market_per_lp_base(
     per_lp_base: i8,
 ) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
 
     let old_per_lp_base = perp_market.amm.per_lp_base;
     msg!(
@@ -2854,7 +3063,19 @@ pub fn handle_update_perp_market_oracle(
     oracle_source: OracleSource,
 ) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
+
     let clock = Clock::get()?;
+
+    OracleMap::validate_oracle_account_info(&ctx.accounts.oracle)?;
+
+    validate!(
+        ctx.accounts.oracle.key == &oracle,
+        ErrorCode::DefaultError,
+        "oracle account info ({:?}) and ix data ({:?}) must match",
+        ctx.accounts.oracle.key,
+        oracle
+    )?;
 
     // Verify oracle is readable
     let OraclePriceData {
@@ -2889,6 +3110,8 @@ pub fn handle_update_perp_market_base_spread(
     base_spread: u32,
 ) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
+
     msg!(
         "perp_market.amm.base_spread: {:?} -> {:?}",
         perp_market.amm.base_spread,
@@ -2927,6 +3150,7 @@ pub fn handle_update_amm_jit_intensity(
     )?;
 
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
 
     msg!(
         "perp_market.amm.amm_jit_intensity: {} -> {}",
@@ -2947,6 +3171,8 @@ pub fn handle_update_perp_market_max_spread(
     max_spread: u32,
 ) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
+
     validate!(
         max_spread >= perp_market.amm.base_spread,
         ErrorCode::DefaultError,
@@ -2979,6 +3205,8 @@ pub fn handle_update_perp_market_step_size_and_tick_size(
     tick_size: u64,
 ) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
+
     validate!(step_size > 0 && tick_size > 0, ErrorCode::DefaultError)?;
     validate!(step_size <= 2000000000, ErrorCode::DefaultError)?; // below i32 max for lp's remainder_base_asset
 
@@ -3007,6 +3235,8 @@ pub fn handle_update_perp_market_min_order_size(
     order_size: u64,
 ) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
+
     validate!(order_size > 0, ErrorCode::DefaultError)?;
 
     msg!(
@@ -3028,6 +3258,8 @@ pub fn handle_update_spot_market_step_size_and_tick_size(
     tick_size: u64,
 ) -> Result<()> {
     let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
+    msg!("spot market {}", spot_market.market_index);
+
     validate!(
         spot_market.market_index == 0 || step_size > 0 && tick_size > 0,
         ErrorCode::DefaultError
@@ -3058,6 +3290,8 @@ pub fn handle_update_spot_market_min_order_size(
     order_size: u64,
 ) -> Result<()> {
     let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
+    msg!("spot market {}", spot_market.market_index);
+
     validate!(
         spot_market.market_index == 0 || order_size > 0,
         ErrorCode::DefaultError
@@ -3082,6 +3316,7 @@ pub fn handle_update_perp_market_max_slippage_ratio(
 ) -> Result<()> {
     validate!(max_slippage_ratio > 0, ErrorCode::DefaultError)?;
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
 
     msg!(
         "perp_market.amm.max_slippage_ratio: {:?} -> {:?}",
@@ -3102,6 +3337,7 @@ pub fn handle_update_perp_market_max_fill_reserve_fraction(
 ) -> Result<()> {
     validate!(max_fill_reserve_fraction > 0, ErrorCode::DefaultError)?;
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
 
     msg!(
         "perp_market.amm.max_fill_reserve_fraction: {:?} -> {:?}",
@@ -3121,6 +3357,7 @@ pub fn handle_update_perp_market_max_open_interest(
     max_open_interest: u128,
 ) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
 
     validate!(
         is_multiple_of_step_size(
@@ -3149,6 +3386,7 @@ pub fn handle_update_perp_market_fee_adjustment(
     fee_adjustment: i16,
 ) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
 
     validate!(
         fee_adjustment.unsigned_abs().cast::<u64>()? <= FEE_ADJUSTMENT_MAX,
@@ -3174,6 +3412,7 @@ pub fn handle_update_perp_market_number_of_users(
     number_of_users_with_base: Option<u32>,
 ) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
 
     if let Some(number_of_users) = number_of_users {
         msg!(
@@ -3213,6 +3452,7 @@ pub fn handle_update_perp_market_fuel(
     fuel_boost_position: Option<u8>,
 ) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
+    msg!("perp market {}", perp_market.market_index);
 
     if let Some(fuel_boost_taker) = fuel_boost_taker {
         msg!(
@@ -3258,6 +3498,7 @@ pub fn handle_update_spot_market_fee_adjustment(
     fee_adjustment: i16,
 ) -> Result<()> {
     let spot = &mut load_mut!(ctx.accounts.spot_market)?;
+    msg!("spot market {}", spot.market_index);
 
     validate!(
         fee_adjustment.unsigned_abs().cast::<u64>()? <= FEE_ADJUSTMENT_MAX,
@@ -3283,8 +3524,10 @@ pub fn handle_update_spot_market_fuel(
     fuel_boost_borrows: Option<u8>,
     fuel_boost_taker: Option<u8>,
     fuel_boost_maker: Option<u8>,
+    fuel_boost_insurance: Option<u8>,
 ) -> Result<()> {
     let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
+    msg!("spot market {}", spot_market.market_index);
 
     if let Some(fuel_boost_taker) = fuel_boost_taker {
         msg!(
@@ -3328,6 +3571,17 @@ pub fn handle_update_spot_market_fuel(
         spot_market.fuel_boost_borrows = fuel_boost_borrows;
     } else {
         msg!("perp_market.fuel_boost_borrows: unchanged");
+    }
+
+    if let Some(fuel_boost_insurance) = fuel_boost_insurance {
+        msg!(
+            "perp_market.fuel_boost_insurance: {:?} -> {:?}",
+            spot_market.fuel_boost_insurance,
+            fuel_boost_insurance
+        );
+        spot_market.fuel_boost_insurance = fuel_boost_insurance;
+    } else {
+        msg!("perp_market.fuel_boost_insurance: unchanged");
     }
 
     Ok(())
@@ -3481,6 +3735,7 @@ pub fn handle_initialize_prelaunch_oracle(
     params: PrelaunchOracleParams,
 ) -> Result<()> {
     let mut oracle = ctx.accounts.prelaunch_oracle.load_init()?;
+    msg!("perp market {}", params.perp_market_index);
 
     oracle.perp_market_index = params.perp_market_index;
     if let Some(price) = params.price {
@@ -3501,6 +3756,8 @@ pub fn handle_update_prelaunch_oracle_params(
 ) -> Result<()> {
     let mut oracle = ctx.accounts.prelaunch_oracle.load_mut()?;
     let mut perp_market = ctx.accounts.perp_market.load_mut()?;
+    msg!("perp market {}", perp_market.market_index);
+
     let now = Clock::get()?.unix_timestamp;
 
     if let Some(price) = params.price {
@@ -3538,6 +3795,7 @@ pub fn handle_delete_prelaunch_oracle(
     _perp_market_index: u16,
 ) -> Result<()> {
     let perp_market = ctx.accounts.perp_market.load()?;
+    msg!("perp market {}", perp_market.market_index);
 
     validate!(
         perp_market.amm.oracle != ctx.accounts.prelaunch_oracle.key(),
@@ -3979,6 +4237,19 @@ pub struct AdminDisableBidAskTwapUpdate<'info> {
         has_one = admin
     )]
     pub state: Box<Account<'info, State>>,
+    #[account(mut)]
+    pub user_stats: AccountLoader<'info, UserStats>,
+}
+
+#[derive(Accounts)]
+pub struct InitUserFuel<'info> {
+    #[account(
+        address = fuel_airdrop_wallet::id()
+    )]
+    pub admin: Signer<'info>, // todo
+    pub state: Box<Account<'info, State>>,
+    #[account(mut)]
+    pub user: AccountLoader<'info, User>,
     #[account(mut)]
     pub user_stats: AccountLoader<'info, UserStats>,
 }