+++ programs/drift/src/state/oracle_map.rs
@@ -22,6 +22,7 @@ use std::slice::Iter;
 use super::state::ValidityGuardRails;
 use crate::math::safe_unwrap::SafeUnwrap;
 use crate::state::traits::Size;
+use crate::validate;
 
 pub const PYTH_1M_IDS: [Pubkey; 2] = [bonk_oracle::id(), pepe_oracle::id()];
 pub const PYTH_PULL_1M_IDS: [Pubkey; 2] = [bonk_pull_oracle::id(), pepe_pull_oracle::id()];
@@ -418,6 +419,18 @@ impl<'a> OracleMap<'a> {
                 has_sufficient_number_of_data_points: true,
             },
         })
+    }
+
+    pub fn validate_oracle_account_info<'c>(account_info: &'c AccountInfo<'a>) -> DriftResult {
+        if *account_info.key == Pubkey::default() {
+            return Ok(());
+        }
+
+        validate!(
+            OracleMap::load_one(account_info, 0, None)?.oracles.len() == 1,
+            ErrorCode::InvalidOracle,
+            "oracle owner not recognizable"
+        )
     }
 }