+++ programs/drift/src/state/perp_market.rs
@@ -579,6 +579,10 @@ impl PerpMarket {
 
         Ok(true)
     }
+
+    pub fn can_sanitize_market_order_auctions(&self) -> DriftResult<bool> {
+        Ok(self.amm.oracle_source != OracleSource::Prelaunch)
+    }
 }
 
 #[cfg(test)]