+++ CHANGELOG.md
@@ -11,8 +11,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - program: track fuel ([#1048](https://github.com/drift-labs/protocol-v2/pull/1048))
 - program: add post multi pyth oracle updates atomic ([#1133](https://github.com/drift-labs/protocol-v2/pull/1133))
+- program: track fuel for if staking ([#1127](https://github.com/drift-labs/protocol-v2/pull/1127))
+- program: validate fee structure ([#1075](https://github.com/drift-labs/protocol-v2/pull/1075))
+- program: check 5 min oracle twap divergence in trigger order ([#1116](https://github.com/drift-labs/protocol-v2/pull/1116))
 
 ### Fixes
+
+- program: more oracle validation in admin fn ([#1082](https://github.com/drift-labs/protocol-v2/pull/1082))
+- program: account for serum already having open order account ([#1077](https://github.com/drift-labs/protocol-v2/pull/1077))
+- program: avoid truncated cast ([#1078](https://github.com/drift-labs/protocol-v2/pull/1078))
+- program: check whitelist token amount ([#1076](https://github.com/drift-labs/protocol-v2/pull/1076))
+- program: program: only let referrer set if number_of_sub_accounts_created is 0 ([#1083](https://github.com/drift-labs/protocol-v2/pull/1083))
 
 ### Breaking